+++ Common/src/main/java/net/merchantpug/apugli/condition/factory/item/OnCooldownCondition.java
@@ -1,6 +1,7 @@
 package net.merchantpug.apugli.condition.factory.item;
 
 import io.github.apace100.calio.data.SerializableData;
+import net.merchantpug.apugli.access.ItemStackAccess;
 import net.merchantpug.apugli.condition.factory.IConditionFactory;
 import net.merchantpug.apugli.platform.Services;
 import net.minecraft.world.entity.player.Player;
@@ -10,7 +11,7 @@ public class OnCooldownCondition implements IConditionFactory<ItemStack> {
 
     @Override
     public boolean check(SerializableData.Instance data, ItemStack stack) {
-        if (Services.PLATFORM.getEntityFromItemStack(stack) instanceof Player player) {
+        if (((ItemStackAccess)(Object)stack).apugli$getEntity() instanceof Player player) {
             return player.getCooldowns().isOnCooldown(stack.getItem());
         }
         return false;