+++ Common/src/main/java/net/merchantpug/apugli/mixin/xplatform/common/ItemStackMixin.java
@@ -1,13 +1,13 @@
 package net.merchantpug.apugli.mixin.xplatform.common;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import net.merchantpug.apugli.access.ItemStackAccess;
 import net.merchantpug.apugli.platform.Services;
 import net.merchantpug.apugli.power.EdibleItemPower;
 import net.merchantpug.apugli.power.factory.ModifyEnchantmentLevelPowerFactory;
 import net.merchantpug.apugli.registry.power.ApugliPowers;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
-import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
@@ -25,7 +25,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import java.util.Optional;
 
 @Mixin(ItemStack.class)
-public abstract class ItemStackMixin {
+public abstract class ItemStackMixin implements ItemStackAccess {
     @Shadow public abstract Item getItem();
 
     @Shadow public abstract CompoundTag getOrCreateTag();
@@ -35,22 +35,30 @@ public abstract class ItemStackMixin {
     @Unique
     private int apugli$previousDamage;
 
+    @Unique
+    public Entity apugli$entity;
+
+    public void apugli$setEntity(Entity entity) { this.apugli$entity = entity; }
+
+    public Entity apugli$getEntity() {
+        return this.apugli$entity;
+    }
     @ModifyReturnValue(method = "copy", at = @At("RETURN"))
-    private ItemStack apugli$copyNewParams(ItemStack original) {
-        Entity holder = Services.PLATFORM.getEntityFromItemStack(original);
+    private ItemStack apugli$copyNewParams(ItemStack newStack) {
+        Entity holder = this.apugli$entity;
         if (holder != null) {
-            if (original.isEmpty()) {
-                original = ModifyEnchantmentLevelPowerFactory.getWorkableEmptyStack(holder);
+            if (newStack.isEmpty()) {
+                newStack = ModifyEnchantmentLevelPowerFactory.getWorkableEmptyStack(holder);
             } else {
-                Services.PLATFORM.setEntityToItemStack(original, holder);
+                ((ItemStackAccess)(Object)newStack).apugli$setEntity(holder);
             }
         }
-        return original;
+        return newStack;
     }
 
     @ModifyVariable(method = "setDamageValue", at = @At(value = "HEAD"), argsOnly = true)
     private int apugli$captureDamageValue(int newDamage) {
-        if (Services.PLATFORM.getEntityFromItemStack((ItemStack)(Object)this) instanceof LivingEntity living) {
+        if (this.apugli$entity instanceof LivingEntity living) {
             CompoundTag tag = this.getOrCreateTag();
             apugli$previousDamage = tag.contains("Damage", Tag.TAG_INT) ? tag.getInt("Damage") : 0;
             int removedDurability = apugli$previousDamage - newDamage;
@@ -66,7 +74,7 @@ public abstract class ItemStackMixin {
 
     @Inject(method = "setDamageValue", at = @At(value = "TAIL"))
     private void apugli$executeDurabilityChangeActions(int newDamage, CallbackInfo ci) {
-        if (!(Services.PLATFORM.getEntityFromItemStack((ItemStack)(Object)this) instanceof LivingEntity living)) return;
+        if (!(this.apugli$entity instanceof LivingEntity living)) return;
         int addedDamage = newDamage - apugli$previousDamage;
         Services.POWER.getPowers(living, ApugliPowers.ACTION_ON_DURABILITY_CHANGE.get()).stream().filter(p -> p.doesApply((ItemStack)(Object)this)).forEach(power -> {
             if (newDamage >= this.getMaxDamage()) {
@@ -82,7 +90,7 @@ public abstract class ItemStackMixin {
     @Inject(method = "getUseAnimation", at = @At("HEAD"), cancellable = true)
     private void apugli$getUseAction(CallbackInfoReturnable<UseAnim> cir) {
         ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
+        if (!(this.apugli$entity instanceof LivingEntity living)) return;
         Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack)).findFirst();
         power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getUseAction().equals(EdibleItemPower.EatAnimation.DRINK) ? UseAnim.DRINK : UseAnim.EAT));
     }
@@ -90,25 +98,9 @@ public abstract class ItemStackMixin {
     @Inject(method = "getUseDuration", at = @At("HEAD"), cancellable = true)
     private void apugli$getMaxUseTime(CallbackInfoReturnable<Integer> cir) {
         ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
+        if (!(this.apugli$entity instanceof LivingEntity living)) return;
         Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack)).findFirst();
         power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getFoodComponent().isFastFood() ? 16 : 32));
-    }
-
-    @Inject(method = "getDrinkingSound", at = @At("HEAD"), cancellable = true)
-    private void apugli$getDrinkSound(CallbackInfoReturnable<SoundEvent> cir) {
-        ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
-        Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack) && p.getSound() != null).findFirst();
-        power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getSound()));
-    }
-
-    @Inject(method = "getEatingSound", at = @At("HEAD"), cancellable = true)
-    private void apugli$getEatSound(CallbackInfoReturnable<SoundEvent> cir) {
-        ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
-        Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack) && p.getSound() != null).findFirst();
-        power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getSound()));
     }
 
 }