+++ Fabric/src/main/java/net/merchantpug/apugli/mixin/fabric/common/ItemStackMixin.java
@@ -9,7 +9,6 @@ import net.merchantpug.apugli.power.EdibleItemPower;
 import net.merchantpug.apugli.registry.power.ApugliPowers;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BucketItem;
@@ -22,7 +21,6 @@ import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -36,19 +34,10 @@ public abstract class ItemStackMixin implements ItemStackAccess {
 
     @Shadow public abstract ItemStack copy();
 
-    @Unique
-    public Entity apugli$entity;
-
-    public void apugli$setEntity(Entity entity) { this.apugli$entity = entity; }
-
-    public Entity apugli$getEntity() {
-        return this.apugli$entity;
-    }
-
     @Inject(method = "use", at = @At("HEAD"), cancellable = true)
     private void apugliuseEdibleItem(Level world, Player user, InteractionHand hand, CallbackInfoReturnable<InteractionResultHolder<ItemStack>> cir) {
         ItemStack stack = (ItemStack)(Object)this;
-        Optional<EdibleItemPower> power = PowerHolderComponent.getPowers(((ItemStackAccess)(Object)stack).apugli$getEntity(), EdibleItemPower.class).stream().filter(p -> p.doesApply(world, stack)).findFirst();
+        Optional<EdibleItemPower> power = PowerHolderComponent.getPowers(user, EdibleItemPower.class).stream().filter(p -> p.doesApply(world, stack)).findFirst();
         if (power.isPresent()) {
             ItemStack itemStack = user.getItemInHand(hand);
             if (user.canEat(power.get().getFoodComponent().canAlwaysEat())) {
@@ -66,8 +55,7 @@ public abstract class ItemStackMixin implements ItemStackAccess {
     @Inject(method = "finishUsingItem", at = @At("RETURN"), cancellable = true)
     private void apugli$finishUsingEdibleItem(Level world, LivingEntity user, CallbackInfoReturnable<ItemStack> cir) {
         ItemStack stack = (ItemStack)(Object)this;
-        if (!(((ItemStackAccess)(Object)stack).apugli$getEntity() instanceof LivingEntity living)) return;
-        Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(world, stack)).findFirst();
+        Optional<EdibleItemPower> power = Services.POWER.getPowers(user, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(world, stack)).findFirst();
         if (power.isPresent()) {
             EdibleItemPower.executeEntityActions(user, stack);
             ItemStack newStack = this.copy();